Revise option parsing and add setup script option
Correctly impliment parsing of command line options with a while loop
c.f. https://stackoverflow.com/a/14203146/8931942
Previously had used a for loop, but the case for unkown options was
catching valid options that were being passed later and not treating
them correctly. The while loop implimentation correctly catches and
applies all valid options.

Also add the CLI option --no-bashrc-append which sets the default to be
to write the setup information to a setup script rather than append it
to the user's ~/.bashrc